Fix outlook event invite description with links
Fixed Outlook event invite descriptions to preserve links after
sanitization by applying the prepareCalendarDescription function
